backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
werkzeug==3.1.3
anthropic==0.49.0

# Google integrations
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.164.0

# ML
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.4
